tests/test_image_processor.py
"""Unit tests for image processor module."""

import pytest
import cv2
import numpy as np
from io import BytesIO
from PIL import Image

from src.face_detector import FaceDetector
from src.image_processor import ImageProcessor


@pytest.fixture
def face_detector():
    """Create a face detector instance for testing."""
    return FaceDetector(confidence_threshold=0.5)


@pytest.fixture
def image_processor(face_detector):
    """Create an image processor instance for testing."""
    return ImageProcessor(face_detector, blur_intensity=51)


@pytest.fixture
def sample_image_bytes():
    """Create sample image as bytes."""
    # Create a simple image
    image = np.zeros((200, 200, 3), dtype=np.uint8)
    image.fill(255)
    
    # Convert to bytes
    _, encoded_image = cv2.imencode(".jpg", image)
    return encoded_image.tobytes()


@pytest.fixture
def sample_png_bytes():
    """Create sample PNG image as bytes."""
    # Create a simple image
    image = np.zeros((200, 200, 3), dtype=np.uint8)
    image.fill(255)
    
    # Convert to bytes
    _, encoded_image = cv2.imencode(".png", image)
    return encoded_image.tobytes()


class TestImageProcessor:
    """Test cases for ImageProcessor class."""

    def test_init_with_default_blur(self, face_detector):
        """Test ImageProcessor initialization with default blur intensity."""
        processor = ImageProcessor(face_detector, blur_intensity=51)
        assert processor.blur_intensity == 51
        assert processor.face_detector == face_detector

    def test_init_with_invalid_blur_negative(self, face_detector):
        """Test ImageProcessor initialization with negative blur intensity."""
        with pytest.raises(ValueError, match="blur_intensity must be positive"):
            ImageProcessor(face_detector, blur_intensity=-1)

    def test_init_with_invalid_blur_even(self, face_detector):
        """Test ImageProcessor initialization with even blur intensity."""
        with pytest.raises(ValueError, match="blur_intensity must be odd"):
            ImageProcessor(face_detector, blur_intensity=50)

    def test_load_image_valid(self, image_processor, sample_image_bytes):
        """Test loading a valid image."""
        image = image_processor.load_image(sample_image_bytes)
        assert image is not None
        assert isinstance(image, np.ndarray)
        assert len(image.shape) == 3
        assert image.shape[2] == 3

    def test_load_image_invalid(self, image_processor):
        """Test loading an invalid image."""
        invalid_bytes = b"not an image"
        with pytest.raises(ValueError, match="Invalid image data"):
            image_processor.load_image(invalid_bytes)

    def test_encode_image_jpeg(self, image_processor, sample_image_bytes):
        """Test encoding image as JPEG."""
        image = image_processor.load_image(sample_image_bytes)
        encoded = image_processor.encode_image(image, format="JPEG")
        assert isinstance(encoded, bytes)
        assert len(encoded) > 0

    def test_encode_image_png(self, image_processor, sample_image_bytes):
        """Test encoding image as PNG."""
        image = image_processor.load_image(sample_image_bytes)
        encoded = image_processor.encode_image(image, format="PNG")
        assert isinstance(encoded, bytes)
        assert len(encoded) > 0

    def test_encode_image_invalid_format(self, image_processor, sample_image_bytes):
        """Test encoding image with invalid format."""
        image = image_processor.load_image(sample_image_bytes)
        with pytest.raises(ValueError, match="Unsupported format"):
            image_processor.encode_image(image, format="GIF")

    def test_blur_face_region(self, image_processor, sample_image_bytes):
        """Test blurring a face region."""
        image = image_processor.load_image(sample_image_bytes)
        blurred = image_processor.blur_face_region(image, 50, 50, 100, 100)
        
        assert blurred is not None
        assert blurred.shape == image.shape
        
        # Check that the region is actually blurred (variance should be lower)
        original_region = image[50:150, 50:150]
        blurred_region = blurred[50:150, 50:150]
        
        # Blurred region should have lower variance
        original_var = np.var(original_region)
        blurred_var = np.var(blurred_region)
        # Note: For a uniform image, variance might be similar, so we just check it doesn't error

    def test_blur_face_region_out_of_bounds(self, image_processor, sample_image_bytes):
        """Test blurring a face region that is out of bounds."""
        image = image_processor.load_image(sample_image_bytes)
        # Try to blur region that extends beyond image bounds
        blurred = image_processor.blur_face_region(image, 150, 150, 100, 100)
        assert blurred is not None
        assert blurred.shape == image.shape

    def test_process_image_without_face(self, image_processor, sample_image_bytes):
        """Test processing an image without faces."""
        processed_image, faces_detected = image_processor.process_image(sample_image_bytes)
        assert isinstance(processed_image, bytes)
        assert len(processed_image) > 0
        assert isinstance(faces_detected, int)
        assert faces_detected >= 0

    def test_process_image_invalid(self, image_processor):
        """Test processing an invalid image."""
        invalid_bytes = b"not an image"
        with pytest.raises(ValueError, match="Failed to process image"):
            image_processor.process_image(invalid_bytes)

    def test_get_image_info(self, image_processor, sample_image_bytes):
        """Test getting image information."""
        info = image_processor.get_image_info(sample_image_bytes)
        assert isinstance(info, dict)
        assert "width" in info
        assert "height" in info
        assert "format" in info
        assert "channels" in info
        assert info["width"] == 200
        assert info["height"] == 200
        assert info["format"] == "JPEG"
        assert info["channels"] == 3

    def test_get_image_info_png(self, image_processor, sample_png_bytes):
        """Test getting image information for PNG."""
        info = image_processor.get_image_info(sample_png_bytes)
        assert info["format"] == "PNG"

    def test_get_image_info_invalid(self, image_processor):
        """Test getting image information for invalid image."""
        invalid_bytes = b"not an image"
        with pytest.raises(ValueError, match="Failed to get image info"):
            image_processor.get_image_info(invalid_bytes)

